Moved all module enums to a central class; moved 'none' states to the first enum element
