+++ robot/src/main/java/us/ilite/robot/controller/TestController.java
@@ -17,9 +17,8 @@ import us.ilite.common.types.sensor.EGyro;
 import us.ilite.robot.Enums;
 import us.ilite.robot.Robot;
 import us.ilite.robot.modules.DJSpinnerModule;
-import us.ilite.robot.modules.FlywheelModule;
 import us.ilite.robot.modules.Limelight;
-import us.ilite.robot.modules.PowerCellModule;
+import static us.ilite.robot.Enums.*;
 
 import static us.ilite.common.config.InputMap.OPERATOR.FIRE_POWER_CELLS;
 import static us.ilite.common.types.EPowerCellData.*;
@@ -39,8 +38,8 @@ public class TestController extends BaseManualController {
     private double mLimelightZoomThreshold = 7.0;
     private double mStartTime;
 
-    private PowerCellModule.EIntakeState mIntakeState;
-    private PowerCellModule.EArmState mArmState;
+    private EIntakeState mIntakeState;
+    private EArmState mArmState;
     private double mPreviousTime;
     private double mGoalBeamCountBroken = 0;
     private boolean crossedEntry = false;
@@ -201,7 +200,7 @@ public class TestController extends BaseManualController {
 
     protected void updatePowerCells(double pNow) {
         // Default to none
-        db.powercell.set(INTAKE_STATE, PowerCellModule.EArmState.NONE);
+        db.powercell.set(INTAKE_STATE, EArmState.NONE);
 
         if (db.operatorinput.isSet(InputMap.OPERATOR.INTAKE_ACTIVATE) || flywheelinput.isSet(InputMap.FLYWHEEL.BASIC_INTAKE)) {
             setIntakeArmEnabled(pNow, true);
@@ -214,7 +213,7 @@ public class TestController extends BaseManualController {
 //            }
 
         } else if (db.operatorinput.isSet(InputMap.OPERATOR.INTAKE_REVERSE) || flywheelinput.isSet(InputMap.FLYWHEEL.REVERSE_INTAKE)) {
-            db.powercell.set(INTAKE_STATE, PowerCellModule.EArmState.STOW);
+            db.powercell.set(INTAKE_STATE, EArmState.STOW);
             reverseSerializer(pNow);
         } else if (db.operatorinput.isSet(InputMap.OPERATOR.INTAKE_STOW) || flywheelinput.isSet(InputMap.FLYWHEEL.INTAKE_STOW)) {
             setIntakeArmEnabled(pNow, false);
@@ -222,7 +221,7 @@ public class TestController extends BaseManualController {
         } else {
             // TODO - only enable once we have set the hold gains
 //            db.powercell.set(INTAKE_STATE, PowerCellModule.EArmState.HOLD);
-            db.powercell.set(INTAKE_STATE, PowerCellModule.EArmState.NONE);
+            db.powercell.set(INTAKE_STATE, EArmState.NONE);
             db.powercell.set(DESIRED_INTAKE_VELOCITY_FT_S, 0d);
         }
 
@@ -284,22 +283,22 @@ public class TestController extends BaseManualController {
 
     void updateDJBooth(double pNow) {
         if (db.operatorinput.isSet(InputMap.OPERATOR.COLOR_POSITION)) {
-            db.color.set(EColorData.DESIRED_MOTOR_POWER, DJSpinnerModule.EColorWheelState.POSITION.getPower());
+            db.color.set(EColorData.DESIRED_MOTOR_POWER, EColorWheelState.POSITION.getPower());
             int i = (int) db.color.get(EColorData.SENSED_COLOR);
-            DJSpinnerModule.EColorMatch m = DJSpinnerModule.EColorMatch.values()[i];
+            EColorMatch m = EColorMatch.values()[i];
             Color DJ_COLOR = null;
             switch (db.recieveColorFmsRelay()) {
                 case 'B':
-                    DJ_COLOR = DJSpinnerModule.EColorMatch.BLUE.color;
+                    DJ_COLOR = EColorMatch.BLUE.color;
                     break;
                 case 'G':
-                    DJ_COLOR = DJSpinnerModule.EColorMatch.GREEN.color;
+                    DJ_COLOR = EColorMatch.GREEN.color;
                     break;
                 case 'R':
-                    DJ_COLOR = DJSpinnerModule.EColorMatch.RED.color;
+                    DJ_COLOR = EColorMatch.RED.color;
                     break;
                 case 'Y':
-                    DJ_COLOR = DJSpinnerModule.EColorMatch.YELLOW.color;
+                    DJ_COLOR = EColorMatch.YELLOW.color;
                     break;
                 default:
                     DJ_COLOR = null;
@@ -307,20 +306,20 @@ public class TestController extends BaseManualController {
             }
             if (m.color.equals(DJ_COLOR)) {
                 //TODO stop using the module for the desired power
-                db.color.set(EColorData.DESIRED_MOTOR_POWER, DJSpinnerModule.EColorWheelState.OFF.getPower());
+                db.color.set(EColorData.DESIRED_MOTOR_POWER, EColorWheelState.OFF.getPower());
             } else {
-                db.color.set(EColorData.DESIRED_MOTOR_POWER, DJSpinnerModule.EColorWheelState.POSITION.getPower());
-                db.color.set(EColorData.COLOR_WHEEL_MOTOR_STATE, DJSpinnerModule.EColorWheelState.POSITION.ordinal());
+                db.color.set(EColorData.DESIRED_MOTOR_POWER, EColorWheelState.POSITION.getPower());
+                db.color.set(EColorData.COLOR_WHEEL_MOTOR_STATE, EColorWheelState.POSITION.ordinal());
             }
         }
         else if ( db.operatorinput.isSet(InputMap.OPERATOR.COLOR_ROTATION)) {
-            db.color.set(EColorData.DESIRED_MOTOR_POWER, DJSpinnerModule.EColorWheelState.ROTATION.getPower());
+            db.color.set(EColorData.DESIRED_MOTOR_POWER, EColorWheelState.ROTATION.getPower());
             if(db.color.get(EColorData.WHEEL_ROTATION_COUNT) >= DJSpinnerModule.TARGET_ROTATION_COUNT) {
-                db.color.set(EColorData.COLOR_WHEEL_MOTOR_STATE, DJSpinnerModule.EColorWheelState.OFF.ordinal());
-                db.color.set(EColorData.DESIRED_MOTOR_POWER, DJSpinnerModule.EColorWheelState.OFF.getPower());
+                db.color.set(EColorData.COLOR_WHEEL_MOTOR_STATE, EColorWheelState.OFF.ordinal());
+                db.color.set(EColorData.DESIRED_MOTOR_POWER, EColorWheelState.OFF.getPower());
             } else {
-                db.color.set(EColorData.COLOR_WHEEL_MOTOR_STATE, DJSpinnerModule.EColorWheelState.ROTATION.ordinal());
-                db.color.set(EColorData.DESIRED_MOTOR_POWER, DJSpinnerModule.EColorWheelState.ROTATION.getPower());
+                db.color.set(EColorData.COLOR_WHEEL_MOTOR_STATE, EColorWheelState.ROTATION.ordinal());
+                db.color.set(EColorData.DESIRED_MOTOR_POWER, EColorWheelState.ROTATION.getPower());
             }
         }
