--- robot/src/main/java/us/ilite/robot/modules/EDriveState.java
@@ -1,17 +1,0 @@
-package us.ilite.robot.modules;
-
-/**
- * Describes the different states the drivetrain may be in. If we do more Talon-based control (motion profiling, motion magic, etc.)
- * that require us to keep track of our state for configuration purposes, we can add those states here.
- */
-public enum EDriveState {
-    RESET,
-    NORMAL,
-    PATH_FOLLOWING_BASIC,
-    PATH_FOLLOWING_HELIX,
-    PATH_FOLLOWING_RAMSETE,
-    TARGET_ANGLE_LOCK,
-    HOLD,
-    VELOCITY,
-    PERCENT_OUTPUT
-}
