+++ robot/src/main/java/us/ilite/robot/modules/LEDControl.java
@@ -6,9 +6,7 @@ import edu.wpi.first.wpilibj.util.Color;
 import us.ilite.common.config.Settings;
 
 import us.ilite.common.types.EColorData;
-
-import static us.ilite.robot.modules.DJSpinnerModule.EColorMatch;
-import static us.ilite.robot.modules.DJSpinnerModule.EColorWheelState;
+import static us.ilite.robot.Enums.*;
 
 public class LEDControl extends Module {
